♻️ move all non-routing logic out of routes Since the new *.page.vue files aren't actual reusable components I've put them in a separate /templates folder for clarity GraphQl did not change, only moved Ref: #179
