+++ app/src/domain/competency-management/templates/root-competencies-page.vue
