+++ app/src/pages/manage/frameworks/[frameworkId]/index.vue
@@ -5,7 +5,7 @@ defineProps<{
 const ability = useAppAbility();
 </script>
 <template>
-  <template v-if="ability.can('read', 'CompetencyFramework')">
+  <template v-if="ability.can('read', 'Competency')">
     <RootCompetenciesPage :framework-id="frameworkId" />
   </template>
   <template v-else>
