+++ app/src/vite-types/components.d.ts
@@ -7,8 +7,11 @@ declare module '@vue/runtime-core' {
   export interface GlobalComponents {
     AuthAccessDenied: typeof import('./../domain/global/components/auth/auth-access-denied.vue')['default']
     AuthConnectButton: typeof import('./../domain/global/components/auth/auth-connect-button.vue')['default']
+    CompetencyCreatePage: typeof import('./../domain/competency-management/templates/competency-create-page.vue')['default']
+    CompetencyEditPage: typeof import('./../domain/competency-management/templates/competency-edit-page.vue')['default']
     CompetencyFrameworkCreate: typeof import('./../domain/competency-management/components/competency-framework-create.vue')['default']
     CompetencyList: typeof import('./../domain/competency-management/components/competency-list.vue')['default']
+    CompetencyListPage: typeof import('./../domain/competency-management/templates/competency-list-page.vue')['default']
     Dialog: typeof import('@headlessui/vue')['Dialog']
     DialogDescription: typeof import('@headlessui/vue')['DialogDescription']
     DialogTitle: typeof import('@headlessui/vue')['DialogTitle']
@@ -17,6 +20,7 @@ declare module '@vue/runtime-core' {
     FkRadio: typeof import('./../ui/input/fk-radio.vue')['default']
     FkSelect: typeof import('./../ui/select/fk-select.vue')['default']
     FrameworkList: typeof import('./../domain/competency-management/components/framework-list.vue')['default']
+    FrameworkListPage: typeof import('./../domain/competency-management/templates/framework-list-page.vue')['default']
     LanguageSelect: typeof import('./../domain/global/components/layout/language-select.vue')['default']
     Listbox: typeof import('@headlessui/vue')['Listbox']
     ListboxButton: typeof import('@headlessui/vue')['ListboxButton']
@@ -43,7 +47,8 @@ declare module '@vue/runtime-core' {
     RiMenuLine: typeof import('~icons/ri/menu-line')['default']
     RiPencilLine: typeof import('~icons/ri/pencil-line')['default']
     RiUser3Line: typeof import('~icons/ri/user3-line')['default']
-    RootCompetenciesPage: typeof import('./../domain/competency-management/components/root-competencies-page.vue')['default']
+    RootCompetenciesCreatePage: typeof import('./../domain/competency-management/templates/root-competencies-create-page.vue')['default']
+    RootCompetenciesPage: typeof import('./../domain/competency-management/templates/root-competencies-page.vue')['default']
     RootCompetencyCreate: typeof import('./../domain/competency-management/components/root-competency-create.vue')['default']
     Switch: typeof import('@headlessui/vue')['Switch']
     TransitionChild: typeof import('@headlessui/vue')['TransitionChild']
